more PTV-842 fixes
