--- skip/genbank_upload_full_test.py
@@ -1,156 +1,0 @@
-import unittest
-import time
-import os
-import shutil
-
-try:
-    from ConfigParser import ConfigParser  # py2
-except:
-    from configparser import ConfigParser  # py3
-
-from Workspace.WorkspaceClient import Workspace as workspaceService
-from GenomeFileUtil.GenomeFileUtilImpl import GenomeFileUtil
-from GenomeFileUtil.GenomeFileUtilServer import MethodContext
-from DataFileUtil.DataFileUtilClient import DataFileUtil
-
-
-class GenomeFileUtilTest(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        token = os.environ.get('KB_AUTH_TOKEN', None)
-        # WARNING: don't call any logging methods on the context object,
-        # it'll result in a NoneType error
-        cls.ctx = MethodContext(None)
-        cls.ctx.update({'token': token,
-                        'provenance': [
-                            {'service': 'GenomeFileUtil',
-                             'method': 'please_never_use_it_in_production',
-                             'method_params': []
-                             }],
-                        'authenticated': 1})
-        config_file = os.environ.get('KB_DEPLOYMENT_CONFIG', None)
-        cls.cfg = {}
-        config = ConfigParser()
-        config.read(config_file)
-        for nameval in config.items('GenomeFileUtil'):
-            cls.cfg[nameval[0]] = nameval[1]
-        cls.wsURL = cls.cfg['workspace-url']
-        cls.wsClient = workspaceService(cls.wsURL, token=token)
-        cls.serviceImpl = GenomeFileUtil(cls.cfg)
-
-    @classmethod
-    def tearDownClass(cls):
-        if hasattr(cls, 'wsName'):
-            cls.wsClient.delete_workspace({'workspace': cls.wsName})
-            print('Test workspace was deleted')
-
-    def getWsClient(self):
-        return self.__class__.wsClient
-
-    def getWsName(self):
-        if hasattr(self.__class__, 'wsName'):
-            return self.__class__.wsName
-        suffix = int(time.time() * 1000)
-        wsName = "test_GenomeFileUtil_" + str(suffix)
-        self.getWsClient().create_workspace({'workspace': wsName})
-        self.__class__.wsName = wsName
-        return wsName
-
-    def getImpl(self):
-        return self.__class__.serviceImpl
-
-    def getContext(self):
-        return self.__class__.ctx
-
-    def test_genome_upload_fungi(self):
-        gbk_path = "data/fungi/GCF_000002945.1_ASM294v2_genomic.gbff.gz"
-        ws_obj_name = 'FungiGenome.1'
-        result = self.getImpl().genbank_to_genome(
-            self.getContext(),
-            {
-              'file': {
-                  'path': gbk_path},
-              'workspace_name': self.getWsName(),
-              'genome_name': ws_obj_name,
-              'generate_ids_if_needed': 1
-            })[0]
-        self.assertTrue(int(
-            result['genome_info'][10]['Number of Protein Encoding Genes']) > 0)
-
-    def test_genome_upload_two_files(self):
-        gbk_path = "data/drosophila/small_test.tar.gz"
-        ws_obj_name = 'DrosophilaGenome.1'
-        result = self.getImpl().genbank_to_genome(
-            self.getContext(),
-            {
-              'file': {'path': gbk_path},
-              'workspace_name': self.getWsName(),
-              'genome_name': ws_obj_name,
-              'generate_ids_if_needed': 0
-            })[0]
-        self.assertTrue(int(
-            result['genome_info'][10]['Number of Protein Encoding Genes']) > 0)
-
-    def test_feature_id_duplication_bug(self):
-        gbk_path = "data/duplication.gbff"
-        ws_obj_name = 'BugGenome.1'
-        result = self.getImpl().genbank_to_genome(
-            self.getContext(),
-            {
-                'file': {'path': gbk_path},
-                'workspace_name': self.getWsName(),
-                'genome_name': ws_obj_name,
-                'generate_ids_if_needed': 1
-            })[0]
-        self.assertTrue(int(
-            result['genome_info'][10]['Number of Protein Encoding Genes']) > 0)
-
-    def test_upload_prokka(self):
-        gbk_path = "data/prokka/PROKKA_012345.gbk.gz"
-        ws_obj_name = 'CustomGenome.1'
-        result = self.getImpl().genbank_to_genome(
-            self.getContext(),
-            {
-                'file': {'path': gbk_path},
-                'workspace_name': self.getWsName(),
-                'genome_name': ws_obj_name,
-                'generate_ids_if_needed': 1
-            })[0]
-        self.assertEquals(result['genome_info'][10]['Number contigs'], '1')
-        self.assertEquals(result['genome_info'][10]['Number of Protein Encoding Genes'], '0')
-        self.assertEquals(result['genome_info'][10]['Domain'], 'Unknown')
-
-    def test_ftp_upload_bug(self):
-        gbk_url = "ftp://ftp.ncbi.nlm.nih.gov/genomes/all/GCF/001/589/275/GCF_001589275.1_sce3192.1/GCF_001589275.1_sce3192.1_genomic.gbff.gz"
-        ws_obj_name = 'BugGenome.2'
-        result = self.getImpl().genbank_to_genome(
-            self.getContext(),
-            {
-                'file': {'ftp_url': gbk_url },
-                'workspace_name': self.getWsName(),
-                'genome_name': ws_obj_name,
-                'generate_ids_if_needed': 1
-            })[0]
-        self.assertTrue(int(
-            result['genome_info'][10]['Number of Protein Encoding Genes']) > 0)
-
-    def test_simple_shock_upload(self):
-        ### Test for upload from SHOCK - upload the file to shock first
-        print('attempting upload through shock')
-        gbk_path = "data/e_coli/GCF_000005845.2_ASM584v2_genomic.gbff"
-        data_file_cli = DataFileUtil(os.environ['SDK_CALLBACK_URL'])
-        shutil.copy(gbk_path, self.cfg['scratch'])
-        shock_id = data_file_cli.file_to_shock({
-            'file_path': os.path.join(self.cfg['scratch'], gbk_path.split("/")[-1])
-        })['shock_id']
-        print("Running test")
-        ws_obj_name2 = 'MyGenome.2'
-        result = self.getImpl().genbank_to_genome(self.getContext(), {
-                'file': {'shock_id': shock_id},
-                'workspace_name': self.getWsName(),
-                'genome_name': ws_obj_name2,
-            })[0]
-        self.assertIsNotNone(result['genome_ref'])
-        self.assertTrue(int(
-            result['genome_info'][10]['Number of Protein Encoding Genes']) > 0)
-
